feat: 예약 시간 조회/삭제 로직 개선
- timeId 기준 1:N 조회로 변경
- 삭제 조건 추가

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -7,15 +7,21 @@ import org.springframework.stereotype.Service;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.ReservationTimeRequestDTO;
 import roomescape.dto.ReservationTimeResponseDTO;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 
 @Service
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(
+            ReservationTimeRepository reservationTimeRepository,
+            ReservationRepository reservationRepository
+    ) {
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public ReservationTimeResponseDTO addReservationTime(
@@ -34,6 +40,9 @@ public class ReservationTimeService {
     }
 
     public void deleteReservationTime(Long id) {
+        if (reservationRepository.existByTimeId(id)) {
+            throw new IllegalArgumentException("이미 예약된 시간은 삭제할 수 없습니다.");
+        }
         reservationTimeRepository.delete(id);
     }
 }
